checkpoint on restructuring bwaise module
